% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{journalrank_params_get}
\alias{journalrank_params_get}
\title{Reads the configuration file and returns a \code{journalrank_params} object}
\usage{
journalrank_params_get(file = NULL)
}
\arguments{
\item{file}{Path and name of the file containing the configuration parameters. By default reads the configuration file included with the package}
}
\value{
A \code{journalrank_params} object
}
\description{
Reads the configuration file and returns a \code{journalrank_params} object
}
\examples{
\donttest{
# Read the default configuration parameters file
params <- journalrank_params_get()

# Read a personal configuration file
params <- journalrank_params_get("MyProject/MyConfFile.yml")
}
}
\seealso{
\code{\link{journalrank_params_open}}
}
