% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{journalrank_params_getpath}
\alias{journalrank_params_getpath}
\title{Gets and checks the path for the default YAML params file}
\usage{
journalrank_params_getpath(file = NULL)
}
\arguments{
\item{file}{Path and name of the file to be checked}
}
\value{
The full path of the configuration file
}
\description{
Gets and checks the path for the default YAML params file
}
