% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{journalrank_params_open}
\alias{journalrank_params_open}
\title{Opens the internal configuration file in the default editor, so that the user can change it}
\usage{
journalrank_params_open()
}
\description{
Opens the internal configuration file in the default editor, so that the user can change it
}
