% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{journalrank_params_getpath}
\alias{journalrank_params_getpath}
\title{Get and check the path for the YAML params file}
\usage{
journalrank_params_getpath()
}
\description{
Get and check the path for the YAML params file
}
